server/requirements.txt
Django==5.1
djangorestframework==3.15.2
pytest-django==4.9.0
time-machine==2.15.0